backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pydantic
black              # optional, for formatting
pytest             # optional, for tests
astor
tree_sitter_languages